Cleanup code, improve returned data for V2 API vacuums
