+++ custom_components/xiaomi_cloud_map_extractor/const.py
@@ -171,3 +171,5 @@ CONF_DEFAULT_ROOM_COLORS = [COLOR_ROOM_1, COLOR_ROOM_2, COLOR_ROOM_3, COLOR_ROOM
                             COLOR_ROOM_13, COLOR_ROOM_14, COLOR_ROOM_15, COLOR_ROOM_16]
 
 V2_MODELS = ['viomi.vacuum.v6']
+
+MM = 50
