+++ custom_components/xiaomi_cloud_map_extractor/map_data_parser.py
@@ -1,312 +1,17 @@
 import logging
-from typing import List, Dict, Optional
-from struct import unpack_from
 
 from .const import *
 from .image_handler import ImageHandler
+from .map_data import MapData
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class ParsingBuffer:
-    def __init__(self, name, data:bytes, start_offs, length):
-        self._name = name
-        self._data = data
-        self._offs = start_offs
-        self._length = length
-
-    def set_name(self, name):
-        self._name = name
-        _LOGGER.debug('SECTION %s: offset 0x%x', self._name, self._offs)
-
-    def skip(self, field, n):
-        if self._length < n:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += n
-        self._length -= n
-
-    def get_uint8(self, field):
-        if self._length < 1:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += 1
-        self._length -= 1
-        return self._data[self._offs - 1]
-
-    def get_uint16(self, field):
-        if self._length < 2:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += 2
-        self._length -= 2
-        return unpack_from('<H', self._data, self._offs - 2)[0]
-
-    def get_uint32(self, field):
-        if self._length < 4:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += 4
-        self._length -= 4
-        return unpack_from('<L', self._data, self._offs - 4)[0]
-
-    def get_float32(self, field):
-        if self._length < 4:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += 4
-        self._length -= 4
-        return unpack_from('<f', self._data, self._offs - 4)[0]
-
-    def get_string_len8(self, field):
-        n = self.get_uint8(field + '.len')
-        if self._length < n:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        self._offs += n
-        self._length -= n
-        return self._data[self._offs - n:self._offs].decode('UTF-8')
-
-    def peek_uint32(self, field):
-        if self._length < 4:
-            raise ValueError(f"error parsing {self._name}.{field} at offset {self._offs:#x}: buffer underrun")
-        return unpack_from('<L', self._data, self._offs)[0]
-
-
 class MapDataParser:
-    CHARGER = 1
-    IMAGE = 2
-    PATH = 3
-    GOTO_PATH = 4
-    GOTO_PREDICTED_PATH = 5
-    CURRENTLY_CLEANED_ZONES = 6
-    GOTO_TARGET = 7
-    ROBOT_POSITION = 8
-    NO_GO_AREAS = 9
-    VIRTUAL_WALLS = 10
-    BLOCKS = 11
-    NO_MOPPING_AREAS = 12
-    OBSTACLES = 13
-    IGNORED_OBSTACLES = 14
-    OBSTACLES_WITH_PHOTO = 15
-    IGNORED_OBSTACLES_WITH_PHOTO = 16
-    CARPET_MAP = 17
-    DIGEST = 1024
-    MM = 50
-    SIZE = 1024
-    KNOWN_OBSTACLE_TYPES = {
-        0: 'cable',
-        2: 'shoes',
-        3: 'poop',
-        5: 'extension cord',
-        9: 'weighting scale',
-        10: 'clothes'
-    }
 
     @staticmethod
-    def parse(raw: bytes, colors, drawables, texts, sizes, image_config):
-        map_data = MapData()
-        map_header_length = MapDataParser.get_int16(raw, 0x02)
-        map_data.major_version = MapDataParser.get_int16(raw, 0x08)
-        map_data.minor_version = MapDataParser.get_int16(raw, 0x0A)
-        map_data.map_index = MapDataParser.get_int32(raw, 0x0C)
-        map_data.map_sequence = MapDataParser.get_int32(raw, 0x10)
-        block_start_position = map_header_length
-        img_start = None
-        while block_start_position < len(raw):
-            block_header_length = MapDataParser.get_int16(raw, block_start_position + 0x02)
-            header = MapDataParser.get_bytes(raw, block_start_position, block_header_length)
-            block_type = MapDataParser.get_int16(header, 0x00)
-            block_data_length = MapDataParser.get_int32(header, 0x04)
-            block_data_start = block_start_position + block_header_length
-            data = MapDataParser.get_bytes(raw, block_data_start, block_data_length)
-            if block_type == MapDataParser.CHARGER:
-                map_data.charger = MapDataParser.parse_charger(block_start_position, raw)
-            elif block_type == MapDataParser.IMAGE:
-                img_start = block_start_position
-                image, rooms = MapDataParser.parse_image(block_data_length, block_header_length, data, header, colors,
-                                                         image_config)
-                map_data.image = image
-                map_data.rooms = rooms
-            elif block_type == MapDataParser.ROBOT_POSITION:
-                map_data.vacuum_position = MapDataParser.parse_vacuum_position(block_data_length, data)
-            elif block_type == MapDataParser.PATH:
-                map_data.path = MapDataParser.parse_path(block_start_position, header, raw)
-            elif block_type == MapDataParser.GOTO_PATH:
-                map_data.goto_path = MapDataParser.parse_path(block_start_position, header, raw)
-            elif block_type == MapDataParser.GOTO_PREDICTED_PATH:
-                map_data.predicted_path = MapDataParser.parse_path(block_start_position, header, raw)
-            elif block_type == MapDataParser.CURRENTLY_CLEANED_ZONES:
-                map_data.zones = MapDataParser.parse_zones(data, header)
-            elif block_type == MapDataParser.GOTO_TARGET:
-                map_data.goto = MapDataParser.parse_goto_target(data)
-            elif block_type == MapDataParser.DIGEST:
-                map_data.is_valid = True
-            elif block_type == MapDataParser.VIRTUAL_WALLS:
-                map_data.walls = MapDataParser.parse_walls(data, header)
-            elif block_type == MapDataParser.NO_GO_AREAS:
-                map_data.no_go_areas = MapDataParser.parse_area(header, data)
-            elif block_type == MapDataParser.NO_MOPPING_AREAS:
-                map_data.no_mopping_areas = MapDataParser.parse_area(header, data)
-            elif block_type == MapDataParser.OBSTACLES:
-                map_data.obstacles = MapDataParser.parse_obstacles(data, header)
-            elif block_type == MapDataParser.IGNORED_OBSTACLES:
-                map_data.ignored_obstacles = MapDataParser.parse_obstacles(data, header)
-            elif block_type == MapDataParser.OBSTACLES_WITH_PHOTO:
-                map_data.obstacles_with_photo = MapDataParser.parse_obstacles(data, header)
-            elif block_type == MapDataParser.IGNORED_OBSTACLES_WITH_PHOTO:
-                map_data.ignored_obstacles_with_photo = MapDataParser.parse_obstacles(data, header)
-            elif block_type == MapDataParser.BLOCKS:
-                block_pairs = MapDataParser.get_int16(header, 0x08)
-                map_data.blocks = MapDataParser.get_bytes(data, 0, block_pairs)
-            block_start_position = block_start_position + block_data_length + MapDataParser.get_int8(header, 2)
-        if not map_data.image.is_empty:
-            MapDataParser.draw_elements(colors, drawables, sizes, map_data, image_config)
-            if len(map_data.rooms) > 0 and map_data.vacuum_position is not None:
-                map_data.vacuum_room = MapDataParser.get_current_vacuum_room(img_start, raw, map_data.vacuum_position)
-            ImageHandler.rotate(map_data.image)
-            ImageHandler.draw_texts(map_data.image, texts)
-        return map_data
-
-    @staticmethod
-    def get_current_vacuum_room(block_start_position, raw, vacuum_position):
-        block_header_length = MapDataParser.get_int16(raw, block_start_position + 0x02)
-        header = MapDataParser.get_bytes(raw, block_start_position, block_header_length)
-        block_data_length = MapDataParser.get_int32(header, 0x04)
-        block_data_start = block_start_position + block_header_length
-        data = MapDataParser.get_bytes(raw, block_data_start, block_data_length)
-        image_top = MapDataParser.get_int32(header, block_header_length - 16)
-        image_left = MapDataParser.get_int32(header, block_header_length - 12)
-        image_width = MapDataParser.get_int32(header, block_header_length - 4)
-        x = round(vacuum_position.x / MapDataParser.MM - image_left)
-        y = round(vacuum_position.y / MapDataParser.MM - image_top)
-        room = ImageHandler.get_room_at_pixel(data, image_width, x, y)
-        return room
-
-    @staticmethod
-    def parse_image(block_data_length, block_header_length, data, header, colors, image_config):
-        image_size = block_data_length
-        image_top = MapDataParser.get_int32(header, block_header_length - 16)
-        image_left = MapDataParser.get_int32(header, block_header_length - 12)
-        image_height = MapDataParser.get_int32(header, block_header_length - 8)
-        image_width = MapDataParser.get_int32(header, block_header_length - 4)
-        if image_width \
-                - image_width * (image_config[CONF_TRIM][CONF_LEFT] + image_config[CONF_TRIM][CONF_RIGHT]) / 100 \
-                < MINIMAL_IMAGE_WIDTH:
-            image_config[CONF_TRIM][CONF_LEFT] = 0
-            image_config[CONF_TRIM][CONF_RIGHT] = 0
-        if image_height \
-                - image_height * (image_config[CONF_TRIM][CONF_TOP] + image_config[CONF_TRIM][CONF_BOTTOM]) / 100 \
-                < MINIMAL_IMAGE_HEIGHT:
-            image_config[CONF_TRIM][CONF_TOP] = 0
-            image_config[CONF_TRIM][CONF_BOTTOM] = 0
-        image, rooms = ImageHandler.parse(data, image_width, image_height, colors, image_config)
-        for number, room in rooms.items():
-            rooms[number] = Zone((room[0] + image_left) * MapDataParser.MM,
-                                 (room[1] + image_top) * MapDataParser.MM,
-                                 (room[2] + image_left) * MapDataParser.MM,
-                                 (room[3] + image_top) * MapDataParser.MM)
-        return ImageData(image_size,
-                         image_top,
-                         image_left,
-                         image_height,
-                         image_width,
-                         image_config,
-                         image), rooms
-
-    @staticmethod
-    def parse_goto_target(data):
-        x = MapDataParser.get_int16(data, 0x00)
-        y = MapDataParser.get_int16(data, 0x02)
-        return Point(x, y)
-
-    @staticmethod
-    def parse_vacuum_position(block_data_length, data):
-        x = MapDataParser.get_int32(data, 0x00)
-        y = MapDataParser.get_int32(data, 0x04)
-        a = None
-        if block_data_length > 8:
-            a = MapDataParser.get_int32(data, 0x08)
-        return Point(x, y, a)
-
-    @staticmethod
-    def parse_charger(block_start_position, raw):
-        x = MapDataParser.get_int32(raw, block_start_position + 0x08)
-        y = MapDataParser.get_int32(raw, block_start_position + 0x0C)
-        return Point(x, y)
-
-    @staticmethod
-    def parse_walls(data, header):
-        wall_pairs = MapDataParser.get_int16(header, 0x08)
-        walls = []
-        for wall_start in range(0, wall_pairs * 8, 8):
-            x0 = MapDataParser.get_int16(data, wall_start + 0)
-            y0 = MapDataParser.get_int16(data, wall_start + 2)
-            x1 = MapDataParser.get_int16(data, wall_start + 4)
-            y1 = MapDataParser.get_int16(data, wall_start + 6)
-            walls.append(Wall(x0, y0, x1, y1))
-        return walls
-
-    @staticmethod
-    def parse_obstacles(data, header):
-        obstacle_pairs = MapDataParser.get_int16(header, 0x08)
-        obstacles = []
-        if obstacle_pairs == 0:
-            return obstacles
-        obstacle_size = int(len(data) / obstacle_pairs)
-        for obstacle_start in range(0, obstacle_pairs * obstacle_size, obstacle_size):
-            x = MapDataParser.get_int16(data, obstacle_start + 0)
-            y = MapDataParser.get_int16(data, obstacle_start + 2)
-            details = {}
-            if obstacle_size >= 6:
-                details[ATTR_TYPE] = MapDataParser.get_int16(data, obstacle_start + 4)
-                if details[ATTR_TYPE] in MapDataParser.KNOWN_OBSTACLE_TYPES:
-                    details[ATTR_DESCRIPTION] = MapDataParser.KNOWN_OBSTACLE_TYPES[details[ATTR_TYPE]]
-                if obstacle_size >= 10:
-                    u1 = MapDataParser.get_int16(data, obstacle_start + 6)
-                    u2 = MapDataParser.get_int16(data, obstacle_start + 8)
-                    details[ATTR_CONFIDENCE_LEVEL] = 0 if u2 == 0 else u1 * 10.0 / u2
-                    if obstacle_size == 28 and (data[obstacle_start + 12] & 0xFF) > 0:
-                        txt = MapDataParser.get_bytes(data, obstacle_start + 12, 16)
-                        details[ATTR_PHOTO_NAME] = txt.decode('ascii')
-            obstacles.append(Obstacle(x, y, details))
-        return obstacles
-
-    @staticmethod
-    def parse_zones(data, header):
-        zone_pairs = MapDataParser.get_int16(header, 0x08)
-        zones = []
-        for zone_start in range(0, zone_pairs * 8, 8):
-            x0 = MapDataParser.get_int16(data, zone_start + 0)
-            y0 = MapDataParser.get_int16(data, zone_start + 2)
-            x1 = MapDataParser.get_int16(data, zone_start + 4)
-            y1 = MapDataParser.get_int16(data, zone_start + 6)
-            zones.append(Zone(x0, y0, x1, y1))
-        return zones
-
-    @staticmethod
-    def parse_path(block_start_position, header, raw):
-        path_points = []
-        end_pos = MapDataParser.get_int32(header, 0x04)
-        point_length = MapDataParser.get_int32(header, 0x08)
-        point_size = MapDataParser.get_int32(header, 0x0C)
-        angle = MapDataParser.get_int32(header, 0x10)
-        start_pos = block_start_position + 0x14
-        for pos in range(start_pos, start_pos + end_pos, 4):
-            x = MapDataParser.get_int16(raw, pos)
-            y = MapDataParser.get_int16(raw, pos + 2)
-            path_points.append(Point(x, y))
-        return Path(point_length, point_size, angle, path_points)
-
-    @staticmethod
-    def parse_area(header, data):
-        area_pairs = MapDataParser.get_int16(header, 0x08)
-        areas = []
-        for area_start in range(0, area_pairs * 16, 16):
-            x0 = MapDataParser.get_int16(data, area_start + 0)
-            y0 = MapDataParser.get_int16(data, area_start + 2)
-            x1 = MapDataParser.get_int16(data, area_start + 4)
-            y1 = MapDataParser.get_int16(data, area_start + 6)
-            x2 = MapDataParser.get_int16(data, area_start + 8)
-            y2 = MapDataParser.get_int16(data, area_start + 10)
-            x3 = MapDataParser.get_int16(data, area_start + 12)
-            y3 = MapDataParser.get_int16(data, area_start + 14)
-            areas.append(Area(x0, y0, x1, y1, x2, y2, x3, y3))
-        return areas
+    def parse(raw: bytes, colors, drawables, texts, sizes, image_config) -> MapData:
+        pass
 
     @staticmethod
     def draw_elements(colors, drawables, sizes, map_data, image_config):
@@ -339,461 +44,3 @@ class MapDataParser:
                 ImageHandler.draw_walls(map_data.image, map_data.walls, colors)
             if DRAWABLE_ZONES == drawable and map_data.zones is not None:
                 ImageHandler.draw_zones(map_data.image, map_data.zones, colors)
-
-    @staticmethod
-    def get_bytes(data: bytes, start_index: int, size: int):
-        return data[start_index:  start_index + size]
-
-    @staticmethod
-    def get_int8(data: bytes, address: int):
-        return data[address] & 0xFF
-
-    @staticmethod
-    def get_int16(data: bytes, address: int):
-        return \
-            ((data[address + 0] << 0) & 0xFF) | \
-            ((data[address + 1] << 8) & 0xFFFF)
-
-    @staticmethod
-    def get_int32(data: bytes, address: int):
-        return \
-            ((data[address + 0] << 0) & 0xFF) | \
-            ((data[address + 1] << 8) & 0xFFFF) | \
-            ((data[address + 2] << 16) & 0xFFFFFF) | \
-            ((data[address + 3] << 24) & 0xFFFFFFFF)
-
-
-class ViomiMapDataParser:
-    POSITION_UNKNOWN = 1100
-
-    @staticmethod
-    def parse(raw: bytes, colors, drawables, texts, sizes, image_config):
-        map_data = MapData()
-        buf = ParsingBuffer('header', raw, 0, len(raw))
-        featureFlags = buf.get_uint32('featureFlags')
-        map_id = buf.peek_uint32('map_id')
-        _LOGGER.debug('featureFlags: 0x%x, map_id: %d', featureFlags,  map_id)
-
-        if featureFlags & 0x00000001 != 0:
-            ViomiMapDataParser.parse_section(buf, 'robot_status', map_id)
-            buf.skip('unknown1', 0x28)
-
-        if featureFlags & 0x00000002 != 0:
-            ViomiMapDataParser.parse_section(buf, 'image', map_id)
-            map_data.image, map_data.rooms = ViomiMapDataParser.parse_image(buf, colors, image_config)
-
-        if featureFlags & 0x00000004 != 0:
-            ViomiMapDataParser.parse_section(buf, 'history', map_id)
-            map_data.path = ViomiMapDataParser.parse_history(buf)
-
-        if featureFlags & 0x00000008 != 0:
-            ViomiMapDataParser.parse_section(buf, 'charge_station', map_id)
-            map_data.charger = ViomiMapDataParser.parse_position(buf, 'pos')
-            foo = buf.get_float32('foo')
-            _LOGGER.debug('pos: %s, foo: %f', map_data.charger, foo)
-
-        if featureFlags & 0x00000010 != 0:
-            ViomiMapDataParser.parse_section(buf, 'restricted_areas', map_id)
-            map_data.walls, map_data.no_go_areas = ViomiMapDataParser.parse_restricted_areas(buf)
-
-        if featureFlags & 0x00000020 != 0:
-            ViomiMapDataParser.parse_section(buf, 'cleaning_areas', map_id)
-            map_data.zones = ViomiMapDataParser.parse_cleaning_areas(buf)
-
-        if featureFlags & 0x00000040 != 0:
-            ViomiMapDataParser.parse_section(buf, 'navigate', map_id)
-            buf.skip('unknown1', 4)
-            map_data.goto = ViomiMapDataParser.parse_position(buf, 'pos')
-            foo = buf.get_float32('foo')
-            _LOGGER.debug('pos: %s, foo: %f', map_data.goto, foo)
-
-        if featureFlags & 0x00000080 != 0:
-            ViomiMapDataParser.parse_section(buf, 'realtime', map_id)
-            buf.skip('unknown1', 5)
-            map_data.vacuum_position = ViomiMapDataParser.parse_position(buf, 'pos')
-            foo = buf.get_float32('foo')
-            _LOGGER.debug('pos: %s, foo: %f', map_data.vacuum_position, foo)
-
-        if featureFlags & 0x00000800 != 0:
-            ViomiMapDataParser.parse_section(buf, 'unknown1', map_id)
-            buf.skip('unknown1', 4)
-
-        if featureFlags & 0x00001000 != 0:
-            ViomiMapDataParser.parse_section(buf, 'rooms', map_id)
-            ViomiMapDataParser.parse_rooms(buf)
-
-        if featureFlags & 0x00002000 != 0:
-            ViomiMapDataParser.parse_section(buf, 'unknown2', map_id)
-            ViomiMapDataParser.parse_unknown_section(buf)
-
-        if featureFlags & 0x00004000 != 0:
-            ViomiMapDataParser.parse_section(buf, 'unknown3', map_id)
-            ViomiMapDataParser.parse_unknown_section(buf)
-
-        if not map_data.image.is_empty:
-            MapDataParser.draw_elements(colors, drawables, sizes, map_data, image_config)
-#            if len(map_data.rooms) > 0 and map_data.vacuum_position is not None:
-#                map_data.vacuum_room = MapDataParser.get_current_vacuum_room(img_start, raw, map_data.vacuum_position)
-            ImageHandler.rotate(map_data.image)
-            ImageHandler.draw_texts(map_data.image, texts)
-        return map_data
-
-    @staticmethod
-    def parse_image(buf, colors, image_config):
-        buf.skip('unknown1', 0x08)
-        image_top = 0
-        image_left = 0
-        image_height = buf.get_uint32('image_height')
-        image_width = buf.get_uint32('image_width')
-        buf.skip('unknown2', 20)
-        image_size = image_height * image_width
-        _LOGGER.debug('width: %d, height: %d', image_width, image_height)
-        if image_width \
-                - image_width * (image_config[CONF_TRIM][CONF_LEFT] + image_config[CONF_TRIM][CONF_RIGHT]) / 100 \
-                < MINIMAL_IMAGE_WIDTH:
-            image_config[CONF_TRIM][CONF_LEFT] = 0
-            image_config[CONF_TRIM][CONF_RIGHT] = 0
-        if image_height \
-                - image_height * (image_config[CONF_TRIM][CONF_TOP] + image_config[CONF_TRIM][CONF_BOTTOM]) / 100 \
-                < MINIMAL_IMAGE_HEIGHT:
-            image_config[CONF_TRIM][CONF_TOP] = 0
-            image_config[CONF_TRIM][CONF_BOTTOM] = 0
-        image, rooms = ImageHandler.parse_viomi(buf, image_width, image_height, colors, image_config)
-        _LOGGER.debug('img: number of rooms: %d, numbers: %s', len(rooms), rooms.keys())
-        for number, room in rooms.items():
-            rooms[number] = Zone((room[0] + image_left) * MapDataParser.MM,
-                                 (room[1] + image_top) * MapDataParser.MM,
-                                 (room[2] + image_left) * MapDataParser.MM,
-                                 (room[3] + image_top) * MapDataParser.MM)
-        return ImageData(image_size,
-                         image_top,
-                         image_left,
-                         image_height,
-                         image_width,
-                         image_config,
-                         image), rooms
-
-    @staticmethod
-    def parse_history(buf):
-        path_points = []
-        buf.skip('unknown1', 4)
-        history_count = buf.get_uint32('history_count')
-        for _ in range(history_count):
-            buf.skip('path.unknown1', 1)
-            path_points.append(ViomiMapDataParser.parse_position(buf, 'path'))
-            #buf.skip('path.unknown2', 9)
-        return Path(len(path_points), 1, 0, path_points)
-
-    @staticmethod
-    def parse_restricted_areas(buf):
-        walls = []
-        areas = []
-        buf.skip('unknown1', 4)
-        area_count = buf.get_uint32('area_count')
-        for _ in range(area_count):
-            buf.skip('restricted.unknown1', 12)
-            p1 = ViomiMapDataParser.parse_position(buf, 'p1')
-            p2 = ViomiMapDataParser.parse_position(buf, 'p2')
-            p3 = ViomiMapDataParser.parse_position(buf, 'p3')
-            p4 = ViomiMapDataParser.parse_position(buf, 'p4')
-            buf.skip('restricted.unknown2', 48)
-            _LOGGER.debug('restricted: %s %s %s %s', p1, p2, p3, p4)
-            if p1 == p2 and p3 == p4:
-                walls.append(Wall(p1.x, p1.y, p3.x, p3.y))
-            else:
-                areas.append(Area(p1.x, p1.y, p2.x, p2.y, p3.x, p3.y, p4.x, p4.y))
-        return walls, areas
-
-    @staticmethod
-    def parse_cleaning_areas(buf):
-        zones = []
-        buf.skip('unknown1', 4)
-        area_count = buf.get_uint32('area_count')
-        for _ in range(area_count):
-            buf.skip('area.unknown1', 12)
-            p1 = ViomiMapDataParser.parse_position(buf, 'p1')
-            p2 = ViomiMapDataParser.parse_position(buf, 'p2')
-            p3 = ViomiMapDataParser.parse_position(buf, 'p3')
-            p4 = ViomiMapDataParser.parse_position(buf, 'p4')
-            buf.skip('area.unknown2', 48)
-            zones.append(Zone(p1.x, p1.y, p3.x, p3.y))
-        return zones
-
-    @staticmethod
-    def parse_rooms(buf):
-        map_name = buf.get_string_len8('map_name')
-        map_arg = buf.get_uint32('map_arg')
-        _LOGGER.debug('map#%d: %s', map_arg, map_name)
-        while map_arg > 1:
-            map_name = buf.get_string_len8('map_name')
-            map_arg = buf.get_uint32('map_arg')
-            _LOGGER.debug('map#%d: %s', map_arg, map_name)
-        room_count = buf.get_uint32('room_count')
-        for _ in range(room_count):
-            room_id = buf.get_uint8('room.id')
-            room_name = buf.get_string_len8('room.name')
-            buf.skip('room.unknown1', 1)
-            room_text_pos = ViomiMapDataParser.parse_position(buf, 'xxx')
-            _LOGGER.debug('room#%d: %s %s', room_id, room_name, room_text_pos)
-        buf.skip('unknown1', 6)
-
-    @staticmethod
-    def parse_section(buf, name, map_id):
-        buf.set_name(name)
-        magic = buf.get_uint32('magic')
-        if magic != map_id:
-            raise ValueError(f"error parsing section {name} at offset {buf._offs - 4:#x}: magic check failed {magic:#x}") #FIXME
-
-    @staticmethod
-    def parse_position(buf, name):
-        x = buf.get_float32(name + '.x')
-        y = buf.get_float32(name + '.y')
-        if x == ViomiMapDataParser.POSITION_UNKNOWN or y == ViomiMapDataParser.POSITION_UNKNOWN:
-            return None
-        return Point(int(1000 * x + 20000), int(1000 * y + 20000))
-
-    @staticmethod
-    def parse_unknown_section(buf):
-        n = buf._data[buf._offs:].find(buf._data[4:8])
-        if n >= 0:
-            buf._offs += n
-            buf._length -= n
-            return True
-        else:
-            return False
-
-
-class MapData:
-    def __init__(self):
-        self.blocks = None
-        self.charger: Optional[Point] = None
-        self.goto: Optional[List[Point]] = None
-        self.goto_path: Optional[List[Point]] = None
-        self.image: Optional[ImageData] = None
-        self.no_go_areas: Optional[List[Area]] = None
-        self.no_mopping_areas: Optional[List[Area]] = None
-        self.obstacles: Optional[List[Obstacle]] = None
-        self.ignored_obstacles: Optional[List[Obstacle]] = None
-        self.obstacles_with_photo: Optional[List[Obstacle]] = None
-        self.ignored_obstacles_with_photo: Optional[List[Obstacle]] = None
-        self.path: Optional[List[Point]] = None
-        self.predicted_path: Optional[List[Point]] = None
-        self.rooms: Optional[Dict[int, Zone]] = None
-        self.vacuum_position: Optional[Point] = None
-        self.vacuum_room: Optional[int] = None
-        self.walls: Optional[List[Wall]] = None
-        self.zones: Optional[List[Zone]] = None
-        self.map_name: Optional[str] = None
-
-    def calibration(self):
-        calibration_points = []
-        for point in [Point(25500, 25500), Point(26500, 25500), Point(26500, 26500)]:
-            img_point = point.to_img(self.image.dimensions).rotated(self.image.dimensions)
-            calibration_points.append({
-                "vacuum": {"x": point.x, "y": point.y},
-                "map": {"x": int(img_point.x), "y": int(img_point.y)}
-            })
-        return calibration_points
-
-
-class Point:
-    def __init__(self, x, y, a=None):
-        self.x = x
-        self.y = y
-        self.a = a
-
-    def __str__(self):
-        if self.a is None:
-            return f"({self.x}, {self.y})"
-        return f"({self.x}, {self.y}, a = {self.a})"
-
-    def __eq__(self, other):
-        return self.x == other.x and self.y == other.y and self.a == other.a
-
-    def as_dict(self):
-        if self.a is None:
-            return {
-                ATTR_X: self.x,
-                ATTR_Y: self.y
-            }
-        return {
-            ATTR_X: self.x,
-            ATTR_Y: self.y,
-            ATTR_A: self.a
-        }
-
-    def to_img(self, image_dimensions):
-        x = self.x / MapDataParser.MM - image_dimensions.left
-        y = self.y / MapDataParser.MM - image_dimensions.top
-        y = image_dimensions.height - y - 1
-        return Point(x * image_dimensions.scale, y * image_dimensions.scale)
-
-    def rotated(self, image_dimensions):
-        alpha = image_dimensions.rotation
-        w = int(image_dimensions.width * image_dimensions.scale)
-        h = int(image_dimensions.height * image_dimensions.scale)
-        x = self.x
-        y = self.y
-        while alpha > 0:
-            tmp = y
-            y = w - x
-            x = tmp
-            tmp = h
-            h = w
-            w = tmp
-            alpha = alpha - 90
-        return Point(x, y)
-
-
-class Obstacle(Point):
-    def __init__(self, x, y, details):
-        super().__init__(x, y)
-        self.details = details
-
-    def as_dict(self):
-        return {**super(Obstacle, self).as_dict(), **self.details}
-
-    def __str__(self):
-        return f"({self.x}, {self.y}, details = {self.details})"
-
-
-class ImageDimensions:
-    def __init__(self, top, left, height, width, scale, rotation):
-        self.top = top
-        self.left = left
-        self.height = height
-        self.width = width
-        self.scale = scale
-        self.rotation = rotation
-
-
-class ImageData:
-    def __init__(self, size, top, left, height, width, image_config, data):
-        trim_left = int(image_config[CONF_TRIM][CONF_LEFT] * width / 100)
-        trim_right = int(image_config[CONF_TRIM][CONF_RIGHT] * width / 100)
-        trim_top = int(image_config[CONF_TRIM][CONF_TOP] * height / 100)
-        trim_bottom = int(image_config[CONF_TRIM][CONF_BOTTOM] * height / 100)
-        scale = image_config[CONF_SCALE]
-        rotation = image_config[CONF_ROTATE]
-        self.size = size
-        self.dimensions = ImageDimensions(top + trim_bottom,
-                                          left + trim_left,
-                                          height - trim_top - trim_bottom,
-                                          width - trim_left - trim_right,
-                                          scale,
-                                          rotation)
-        self.is_empty = height == 0 or width == 0
-        self.data = data
-
-    def as_dict(self):
-        return {
-            ATTR_SIZE: self.size,
-            ATTR_OFFSET_Y: self.dimensions.top,
-            ATTR_OFFSET_X: self.dimensions.left,
-            ATTR_HEIGHT: self.dimensions.height,
-            ATTR_SCALE: self.dimensions.scale,
-            ATTR_ROTATION: self.dimensions.rotation,
-            ATTR_WIDTH: self.dimensions.width
-        }
-
-
-class Path:
-    def __init__(self, point_length, point_size, angle, path: list):
-        self.point_length = point_length
-        self.point_size = point_size
-        self.angle = angle
-        self.path = path
-
-    def as_dict(self):
-        return {
-            ATTR_POINT_LENGTH: self.point_length,
-            ATTR_POINT_SIZE: self.point_size,
-            ATTR_ANGLE: self.angle,
-            ATTR_PATH: self.path
-        }
-
-
-class Zone:
-    def __init__(self, x0, y0, x1, y1):
-        self.x0 = x0
-        self.y0 = y0
-        self.x1 = x1
-        self.y1 = y1
-
-    def __str__(self):
-        return f"[{self.x0}, {self.y0}, {self.x1}, {self.y1}]"
-
-    def as_dict(self):
-        return {
-            ATTR_X0: self.x0,
-            ATTR_Y0: self.y0,
-            ATTR_X1: self.x1,
-            ATTR_Y1: self.y1
-        }
-
-    def as_area(self):
-        return Area(self.x0, self.y0, self.x0, self.y1, self.x1, self.y1, self.x1, self.y0)
-
-
-class Wall:
-    def __init__(self, x0, y0, x1, y1):
-        self.x0 = x0
-        self.y0 = y0
-        self.x1 = x1
-        self.y1 = y1
-
-    def __str__(self):
-        return f"[{self.x0}, {self.y0}, {self.x1}, {self.y1}]"
-
-    def as_dict(self):
-        return {
-            ATTR_X0: self.x0,
-            ATTR_Y0: self.y0,
-            ATTR_X1: self.x1,
-            ATTR_Y1: self.y1
-        }
-
-    def to_img(self, image_dimensions):
-        p0 = Point(self.x0, self.y0).to_img(image_dimensions)
-        p1 = Point(self.x1, self.y1).to_img(image_dimensions)
-        return Wall(p0.x, p0.y, p1.x, p1.y)
-
-    def as_list(self):
-        return [self.x0, self.y0, self.x1, self.y1]
-
-
-class Area:
-    def __init__(self, x0, y0, x1, y1, x2, y2, x3, y3):
-        self.x0 = x0
-        self.y0 = y0
-        self.x1 = x1
-        self.y1 = y1
-        self.x2 = x2
-        self.y2 = y2
-        self.x3 = x3
-        self.y3 = y3
-
-    def __str__(self):
-        return f"[{self.x0}, {self.y0}, {self.x1}, {self.y1}, {self.x2}, {self.y2}, {self.x3}, {self.y3}]"
-
-    def as_dict(self):
-        return {
-            ATTR_X0: self.x0,
-            ATTR_Y0: self.y0,
-            ATTR_X1: self.x1,
-            ATTR_Y1: self.y1,
-            ATTR_X2: self.x2,
-            ATTR_Y2: self.y2,
-            ATTR_X3: self.x3,
-            ATTR_Y3: self.y3
-        }
-
-    def as_list(self):
-        return [self.x0, self.y0, self.x1, self.y1, self.x2, self.y2, self.x3, self.y3]
-
-    def to_img(self, image_dimensions):
-        p0 = Point(self.x0, self.y0).to_img(image_dimensions)
-        p1 = Point(self.x1, self.y1).to_img(image_dimensions)
-        p2 = Point(self.x2, self.y2).to_img(image_dimensions)
-        p3 = Point(self.x3, self.y3).to_img(image_dimensions)
-        return Area(p0.x, p0.y, p1.x, p1.y, p2.x, p2.y, p3.x, p3.y)
